+++ src/onLayout/addParticipantCountContainer.js
@@ -1,0 +1,8 @@
+export default function addParticipantCountContainer() {
+    this.wrap
+        .select('.legend')
+        .append('span')
+        .classed('annote', true)
+        .style('float', 'right')
+        .style('font-style', 'italic');
+}
