+++ src/onResize.js
@@ -1,6 +1,7 @@
-import { svg, set } from 'd3';
-import syncColors from './util/syncColors';
-import addHighlightLegendItem from './util/addHighlightLegendItem';
+import syncColors from './onResize/syncColors';
+import addHighlightLegendItem from './onResize/addHighlightLegendItem';
+import drawTopXaxis from './onResize/drawTopXaxis';
+import addTickClick from './onResize/addTickClick';
 
 export default function onResize() {
     let context = this;
@@ -12,66 +13,10 @@ export default function onResize() {
     if (this.config.highlight) addHighlightLegendItem(this);
 
     //Draw second x-axis at top of chart.
-    let x2Axis = svg
-        .axis()
-        .scale(this.x)
-        .orient('top')
-        .tickFormat(this.xAxis.tickFormat())
-        .innerTickSize(this.xAxis.innerTickSize())
-        .outerTickSize(this.xAxis.outerTickSize())
-        .ticks(this.xAxis.ticks()[0]);
-    let g_x2_axis = this.svg.select('g.x2.axis').attr('class', 'x2 axis linear');
-    g_x2_axis.call(x2Axis);
-    g_x2_axis
-        .select('text.axis-title.top')
-        .attr('transform', 'translate(' + this.raw_width / 2 + ',-' + this.config.margin.top + ')');
-    g_x2_axis.select('.domain').attr({
-        fill: 'none',
-        stroke: '#ccc',
-        'shape-rendering': 'crispEdges'
-    });
-    g_x2_axis.selectAll('.tick line').attr('stroke', '#eee');
+    drawTopXaxis.call(this);
 
     //Draw second chart when y-axis tick label is clicked.
-    this.svg
-        .select('.y.axis')
-        .selectAll('.tick')
-        .style('cursor', 'pointer')
-        .on('click', d => {
-            let csv2 = this.raw_data.filter(di => di[this.config.id_col] === d);
-            this.chart2.wrap.style('display', 'block');
-            this.chart2.draw(csv2);
-            this.chart2.wrap
-                .select('#backButton')
-                .append('strong')
-                .attr('class', 'id-title')
-                .style('margin-left', '1%')
-                .text('Participant: ' + d);
-
-            //Sort listing by sequence.
-            const seq_col = context.config.seq_col;
-            let tableData = this.superRaw
-                .filter(di => di[this.config.id_col] === d)
-                .sort((a, b) => (+a[seq_col] < b[seq_col] ? -1 : 1));
-
-            //Define listing columns.
-            this.table.config.cols = set(
-                this.config.details.map(detail => detail.value_col)
-            ).values();
-            this.table.config.headers = set(
-                this.config.details.map(detail => detail.label)
-            ).values();
-            this.table.wrap.style('display', 'block');
-            this.table.draw(tableData);
-            this.table.wrap.selectAll('th,td').style({
-                'text-align': 'left',
-                'padding-right': '10px'
-            });
-
-            //Hide timelines.
-            this.wrap.style('display', 'none');
-            this.controls.wrap.style('display', 'none');
-        });
+    addTickClick.call(this);
 
     /**-------------------------------------------------------------------------------------------\
       Second chart callbacks.
