+++ src/onResize/addHighlightLegendItem.js
