+++ src/onResize/syncColors.js
