+++ src/polyfills/object-assign.js
